rabbitmq/requirements.txt
pika==1.3.*
grpcio==1.48.*
protobuf==3.20.*
